requirements.txt
boto3==1.34.29
botocore==1.34.29
kafka-python==2.0.2
pandas==2.2.0
pyarrow==15.0.0
requests==2.31.0 